MAINT: optimization by eliminating astype

+++ cheartpy-gmsh/src/cheartpy/gmsh/reader/regions.py
@@ -28,4 +28,4 @@ def import_region_mask[I: np.integer](file: Path, dtype: DType[I] = np.intp) -> 
     """
     # Implementation goes here
     mask = chread_d(file, dtype=dtype).flatten()
-    return {k: np.nonzero(mask == k)[0].astype(dtype) for k in np.unique(mask)}
+    return {k: np.nonzero(mask == k)[0].astype(dtype, copy=False) for k in np.unique(mask)}
